requirements.txt
PySide6
pyttsx3
mss
pillow